requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kubernetes==28.1.0
python-multipart==0.0.6
jinja2==3.1.2
pyyaml==6.0.1
passlib[bcrypt]==1.7.4 